data_viewer/requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.5
redis==4.6.0
